+++ src/lib/validation/at-rules/index.ts
@@ -9,3 +9,5 @@ export * from './font-feature-values';
 export * from './namespace';
 export * from './document';
 export * from './keyframes';
+export * from './when';
+export * from './else';